Findin Max of Float Number

+++ MaximumNumber.cs
@@ -9,7 +9,7 @@ namespace FindMaximumUisngGenerics
     public class MaximumNumber
     {
         //Method for Find Maximum Number
-        public int MaximumIntegerNumber(int firstNumber, int secondNumber, int thirdNumber)
+        public double MaximumIntegerNumber(double firstNumber, double secondNumber, double thirdNumber)
         {
             if(firstNumber.CompareTo(secondNumber) > 0 && firstNumber.CompareTo(thirdNumber) > 0)
             {
